Moved functions into proper compilation unit
