+++ commands/service_board_identify_test.go
